pyproject.toml
[project]
name = "calendar-to-google"
version = "1.0.0"
description = "System tray app to add clipboard text to Google Calendar"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "pillow>=12.0.0",
    "pyperclip>=1.11.0",
    "pystray>=0.19.5",
    "python-dateutil>=2.9.0.post0",
    "keyboard>=0.13.5; sys_platform == 'win32'",
    "customtkinter>=5.2.2",
]

[project.scripts]
calendar-to-google = "calendar_to_google.tray_app:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["calendar_to_google"]
